# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for NR-1 Psychosocial Risk Management & AI Recruitment SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/me:
    get:
      operationId: getCurrentTenant
      tags: [tenants]
      summary: Get current tenant info
      responses:
        "200":
          description: Tenant info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, name, plan, activeModules, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        cnpj:
          type: ["string", "null"]
        plan:
          type: string
          enum: [essencial, gestao, enterprise, profissional]
        activeModules:
          type: array
          items:
            type: string
        privacyThreshold:
          type: integer
        createdAt:
          type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        cnpj:
          type: string
        privacyThreshold:
          type: integer
...